+++ src/main/java/insane96mcp/survivalreimagined/mixin/MobEffectMixin.java
@@ -1,6 +1,6 @@
 package insane96mcp.survivalreimagined.mixin;
 
-import insane96mcp.survivalreimagined.module.hungerhealth.feature.NoHunger;
+import insane96mcp.survivalreimagined.module.misc.feature.Misc;
 import net.minecraft.world.effect.MobEffect;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.Constant;
@@ -11,7 +11,7 @@ public class MobEffectMixin {
 
 	@ModifyConstant(method = "isDurationEffectTick", constant = @Constant(intValue = 25, ordinal = 0))
 	public int onPoisonTickDamage(int constant) {
-		if (NoHunger.isSlowerPoison()) return 80;
+		if (Misc.isSlowerPoison()) return 80;
 		return constant;
 	}
 }